requirements.txt
google-generativeai
Pillow
pandas
python-dotenv
openpyxl